Move dummy model to actual fnet repo so importlib knows where to look
